# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinTrack Multi-Company Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get financial summary for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /companies/{companyId}/reports/by-category:
    get:
      operationId: getReportByCategory
      tags: [reports]
      summary: Get transactions grouped by category
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, deposit]
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /companies/{companyId}/reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly income/expense totals
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Monthly report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTotal"

  # Invoices
...
  /companies/{companyId}/receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipts for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Create a receipt
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /companies/{companyId}/receipts/{id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get a receipt
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
    patch:
      operationId: updateReceipt
      tags: [receipts]
      summary: Update a receipt
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete a receipt
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ReceiptLineItem:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    Receipt:
      type: object
      required: [id, companyId, receiptNumber, vendorName, date, lineItems, subtotal, taxRate, taxAmount, total, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        receiptNumber:
          type: string
        vendorName:
          type: string
        vendorEmail:
          type: ["string", "null"]
        vendorAddress:
          type: ["string", "null"]
        date:
          type: string
          format: date
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptLineItem"
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReceiptInput:
      type: object
      required: [vendorName, date, lineItems, taxRate]
      properties:
        receiptNumber:
          type: string
        vendorName:
          type: string
          minLength: 1
        vendorEmail:
          type: string
        vendorAddress:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: integer
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptLineItem"
        taxRate:
          type: number
          minimum: 0
        notes:
          type: string
        paymentMethod:
          type: string
    ReceiptUpdate:
      type: object
      properties:
        receiptNumber:
          type: string
        vendorName:
          type: string
          minLength: 1
        vendorEmail:
          type: string
        vendorAddress:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: ["integer", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptLineItem"
        taxRate:
          type: number
          minimum: 0
        notes:
          type: string
        paymentMethod:
          type: string
    ReportSummary:
      type: object
      required: [totalIncome, totalExpenses, netBalance, transactionCount, startDate, endDate]
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netBalance:
          type: number
        transactionCount:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    CategoryBreakdown:
      type: object
      required: [categoryId, categoryName, type, total, count, percentage]
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        type:
          type: string
          enum: [expense, deposit]
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
        color:
          type: ["string", "null"]
    MonthlyTotal:
      type: object
      required: [month, year, totalIncome, totalExpenses, netBalance]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netBalance:
          type: number